+++ tests/integration/test_transaction.py
@@ -4354,48 +4354,6 @@ class TestTransaction(unittest.TestCase):
         self.assertNotEqual(None, transaction.paypal_details.debug_id)
         self.assertEqual(transaction.paypal_details.payee_id, "fake-payee-id")
 
-    def test_creating_paypal_transaction_with_payee_id_in_options_params(self):
-        result = Transaction.sale({
-            "amount": TransactionAmounts.Authorize,
-            "payment_method_nonce": Nonces.PayPalOneTimePayment,
-            "paypal_account": {},
-            "options": {
-                "payee_id": "fake-payee-id"
-            }
-        })
-
-        self.assertTrue(result.is_success)
-        transaction = result.transaction
-
-        self.assertEqual(transaction.paypal_details.payer_email, "payer@example.com")
-        self.assertNotEqual(None, re.search(r'PAY-\w+', transaction.paypal_details.payment_id))
-        self.assertNotEqual(None, re.search(r'AUTH-\w+', transaction.paypal_details.authorization_id))
-        self.assertNotEqual(None, transaction.paypal_details.image_url)
-        self.assertNotEqual(None, transaction.paypal_details.debug_id)
-        self.assertEqual(transaction.paypal_details.payee_id, "fake-payee-id")
-
-    def test_creating_paypal_transaction_with_payee_id_in_options_paypal_params(self):
-        result = Transaction.sale({
-            "amount": TransactionAmounts.Authorize,
-            "payment_method_nonce": Nonces.PayPalOneTimePayment,
-            "paypal_account": {},
-            "options": {
-                "paypal": {
-                    "payee_id": "fake-payee-id"
-                }
-            }
-        })
-
-        self.assertTrue(result.is_success)
-        transaction = result.transaction
-
-        self.assertEqual(transaction.paypal_details.payer_email, "payer@example.com")
-        self.assertNotEqual(None, re.search(r'PAY-\w+', transaction.paypal_details.payment_id))
-        self.assertNotEqual(None, re.search(r'AUTH-\w+', transaction.paypal_details.authorization_id))
-        self.assertNotEqual(None, transaction.paypal_details.image_url)
-        self.assertNotEqual(None, transaction.paypal_details.debug_id)
-        self.assertEqual(transaction.paypal_details.payee_id, "fake-payee-id")
-
     def test_creating_paypal_transaction_with_payee_email(self):
         result = Transaction.sale({
             "amount": TransactionAmounts.Authorize,